refactor: 将 AgentModalContent 内联至 agents/index.tsx
AgentModalContent 是 agents 模块的唯一子组件，内联后可简化模块结构，减少文件间接层
